Update facet.js
Allow facets to access the tree instance, as a second argument. By doing so the facet can access any state in the tree, not only cursors defined. This gives more flexibility. Also the facet is able to pass the instance it is running on to other dependencies that might need to operate on the instance.

Overall more decoupled :-)